+++ tests/testthat/test-alanwood.R
@@ -7,7 +7,7 @@ test_that("examples in the article are unchanged", {
   aw_data <- aw_query(lc50$cas[1:3], from = "cas")
   igroup <- sapply(aw_data, function(y) y$subactivity[1])
 
-  expect_is(igroup, "character")
+  expect_type(igroup, "character")
   expect_equal(names(igroup), c("50-29-3", "52-68-6", "55-38-9"))
   expect_equal(unname(igroup), c("organochlorine insecticides",
                                  "phosphonate insecticides",
@@ -40,7 +40,7 @@ test_that("alanwood, invalid input", {
 
   comps <- c("balloon", NA)
   o1 <- aw_query(comps)
-  expect_is(o1, "list")
+  expect_type(o1, "list")
   expect_equal(o1[[1]], NA)
   expect_equal(o1[[2]], NA)
 })
@@ -49,7 +49,7 @@ test_that("alanwood, build_index", {
   skip_on_cran()
   skip_if_not(up, "Alanwood service is down")
 
-  idx <- suppressWarnings(build_aw_idx(verbose = FALSE, force_build = TRUE))
+  idx <- suppressWarnings(build_aw_idx(force_build = TRUE))
   expect_s3_class(idx, "data.frame")
   expect_equal(ncol(idx), 4)
   expect_equal(names(idx), c("names", "links", "linknames", "source"))