+++ tests/testthat/test-cir.R
@@ -3,17 +3,16 @@ test_that("cir_query()", {
   skip_on_cran()
   skip_if_not(up, "CIR server is down")
 
-  expect_equal(cir_query('Triclosan', 'mw', verbose = FALSE)[[1]], 289.5451)
-  expect_equal(cir_query('xxxxxxx', 'mw', verbose = FALSE)[[1]], NA)
-  expect_equal(cir_query("3380-34-5", 'stdinchikey', resolver = 'cas_number', verbose = FALSE)[[1]],
+  expect_equal(cir_query('Triclosan', 'mw')[[1]], 289.5451)
+  expect_equal(cir_query('xxxxxxx', 'mw')[[1]], NA)
+  expect_equal(cir_query("3380-34-5", 'stdinchikey', resolver = 'cas_number')[[1]],
             "InChIKey=XEFQLINVKFYRCS-UHFFFAOYSA-N")
-  expect_true(length(cir_query('Triclosan', 'cas', verbose = FALSE)[[1]]) > 1)
-  expect_message(cir_query("acetic acid", "mw", match = "first"))
-  expect_length(cir_query('Triclosan', 'cas', match = "first", verbose = FALSE)[[1]], 1)
-  expect_length(cir_query(c('Triclosan', 'Aspirin'), 'cas', verbose = FALSE), 2)
+  expect_true(length(cir_query('Triclosan', 'cas')[[1]]) > 1)
+  expect_length(cir_query('Triclosan', 'cas', match = "first")[[1]], 1)
+  expect_length(cir_query(c('Triclosan', 'Aspirin'), 'cas'), 2)
 
-  # skip("I have no clue why this one fails on R CMD check.  It works when run in the console!")
-  expect_equivalent(cir_query('acetic acid', 'mw', match = "first"), c(`acetic acid` = 60.0524))
+  expect_equal(cir_query('acetic acid', 'mw', match = "first"),
+               list(`acetic acid` = 60.0524))
 
 })