+++ tests/testthat/test-srs.R
@@ -12,8 +12,8 @@ test_that("SRS returns correct results", {
 
   expect_true(is.na(a))
   expect_true(is.na(b))
-  expect_is(c, "list")
-  expect_is(c$`50-00-0`, "data.frame")
+  expect_type(c, "list")
+  expect_s3_class(c$`50-00-0`, "data.frame")
   expect_equal(c$`50-00-0`$systematicName, "Formaldehyde")
 
   expect_equal(d$aniline$systematicName, "Benzenamine")