+++ NEWS.md
@@ -1,3 +1,13 @@
+# webchem 1.1.1.9001
+
+## BUG FIXES
+
+* ci_query() can no longer query chemicals by name.
+
+## MINOR IMPROVEMENTS
+
+* webchem functions now default to global options regarding verbose messages.
+
 # webchem 1.1.1.
 
 ## NEW FEATURES